+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## v0.1.1 - ?
 
+- Improve desired state stability:
+  - Sort slices by name during unrolling
+  - Preserve embedded entities order from slice source
+- Fix deletion of embedded entities
 
 ## v0.1.0 - 2025-12-03